config/requirements.txt
Flask==0.10.1
Flask-Assets==0.8
Flask-Misaka==0.3.0
Flask-Script==0.6.7
Flask-WTF==0.9.5
Jinja2==2.7.2
Markdown==2.4
MarkupSafe==0.19
Pillow==2.5.2
Pygments==1.6
Sphinx==1.2.3
WTForms==1.0.5
Werkzeug==0.9.4
argparse==1.2.1
blinker==1.3
coverage==3.7.1
docutils==0.12
flake8==2.2.5
flask-mongoengine==0.7.0
gaenv==0.1.7
google-api-python-client==1.2
httplib2==0.8
itsdangerous==0.24
mccabe==0.2.1
misaka==1.0.2
mongoengine==0.8.7
pep8==1.5.7
pyRFC3339==0.2
pyflakes==0.8.1
pymongo==2.7.2
python-gflags==2.0
pytz==2015.2
requests==2.3.0
webassets==0.8
wsgiref==0.1.2
gunicorn==19.2.1
python-consul==0.3.10

